Clone from VolumeSnapshot source controller

+++ pkg/controller/datavolume/smart-clone-controller.go
@@ -238,7 +238,7 @@ func (r *SmartCloneReconciler) reconcileSnapshot(log logr.Logger, snapshot *snap
 	if err != nil {
 		return reconcile.Result{}, err
 	}
-	newPvc, err := newPvcFromSnapshot(snapshot, targetPvcSpec)
+	newPvc, err := newPvcFromSnapshot(snapshot.Name, snapshot, targetPvcSpec)
 	if err != nil {
 		return reconcile.Result{}, err
 	}
@@ -296,6 +296,11 @@ func (r *SmartCloneReconciler) deleteSnapshot(log logr.Logger, namespace, name s
 		return nil
 	}
 
+	if _, ok := snapshotToDelete.Labels[common.CDIComponentLabel]; !ok {
+		// Not a CDI snapshot, don't delete
+		return nil
+	}
+
 	if err := r.client.Delete(context.TODO(), snapshotToDelete); err != nil {
 		if !k8serrors.IsNotFound(err) {
 			log.Error(err, "error deleting snapshot for smart-clone")
@@ -347,7 +352,8 @@ func (r *SmartCloneReconciler) getTargetPVC(dataVolume *cdiv1.DataVolume) (*core
 	return pvc, nil
 }
 
-func newPvcFromSnapshot(snapshot *snapshotv1.VolumeSnapshot, targetPvcSpec *corev1.PersistentVolumeClaimSpec) (*corev1.PersistentVolumeClaim, error) {
+func newPvcFromSnapshot(name string, snapshot *snapshotv1.VolumeSnapshot, targetPvcSpec *corev1.PersistentVolumeClaimSpec) (*corev1.PersistentVolumeClaim, error) {
+	targetPvcSpecCopy := targetPvcSpec.DeepCopy()
 	restoreSize := snapshot.Status.RestoreSize
 	if restoreSize == nil {
 		return nil, fmt.Errorf("snapshot has no RestoreSize")
@@ -363,13 +369,13 @@ func newPvcFromSnapshot(snapshot *snapshotv1.VolumeSnapshot, targetPvcSpec *core
 		common.CDILabelKey:       common.CDILabelValue,
 		common.CDIComponentLabel: common.SmartClonerCDILabel,
 	}
-	if util.ResolveVolumeMode(targetPvcSpec.VolumeMode) == corev1.PersistentVolumeFilesystem {
+	if util.ResolveVolumeMode(targetPvcSpecCopy.VolumeMode) == corev1.PersistentVolumeFilesystem {
 		labels[common.KubePersistentVolumeFillingUpSuppressLabelKey] = common.KubePersistentVolumeFillingUpSuppressLabelValue
 	}
 
 	target := &corev1.PersistentVolumeClaim{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      snapshot.Name,
+			Name:      name,
 			Namespace: snapshot.Namespace,
 			Labels:    labels,
 			Annotations: map[string]string{
@@ -386,10 +392,10 @@ func newPvcFromSnapshot(snapshot *snapshotv1.VolumeSnapshot, targetPvcSpec *core
 				Kind:     "VolumeSnapshot",
 				APIGroup: &snapshotv1.SchemeGroupVersion.Group,
 			},
-			VolumeMode:       targetPvcSpec.VolumeMode,
-			AccessModes:      targetPvcSpec.AccessModes,
-			StorageClassName: targetPvcSpec.StorageClassName,
-			Resources:        targetPvcSpec.Resources,
+			VolumeMode:       targetPvcSpecCopy.VolumeMode,
+			AccessModes:      targetPvcSpecCopy.AccessModes,
+			StorageClassName: targetPvcSpecCopy.StorageClassName,
+			Resources:        targetPvcSpecCopy.Resources,
 		},
 	}
